==> exploring-spotify-streaming-history/00_eda.ipynb <==
# Running some numbers

More info: https://www.notion.so/tacosdedatos/analyzing-my-spotify-streaming-history-f545fb62e86d49159a383f950f224e55 

import pandas as pd
from datetime import datetime as dt
import json 
import altair as alt
from rich import print 

with open("tracks.json") as file:
    tracks = json.load(file)

## Some questions to answer

- What are the most listened to songs during particular times of day?
    - By hour
    - By time of day (2:00-5:00, 5-8, 8-10, 10-1,1-4,4-7,7-10,10-2)
    - An hour before work, an hour after work, saturday morning, sunday morning, monday morning, payday
- What are the top 5 songs streamed by month?
- Are there any correlations between the songs listened to and the weather?
- Are there any patterns in the artist/album/song combinations that are streamed?
- Are there any songs with unusually long gaps between streams?
    - Average time between streams
- What is the most popular artist/album combination?
    - Top album by artist
        - per plays
        - per songs played
        - weighted average
- What are the most common song pairs?
    - Songs I play together
    - Songs I play in sequence
    - 3-song, 4-song, 5-song streaks
- Metrics of my top five:
    - streaming patterns/frequency
        - time of day
        - time of week
        - time of month
    - by song, album
