==> exploring-spotify-streaming-history/02_getting_data.ipynb <==
import json

import pandas as pd
from rich import print 
from rich.progress import track as track_progress

import spotipy
from spotipy.oauth2 import SpotifyClientCredentials
from dotenv import dotenv_values

config = dotenv_values("../.env")

client_id = config.get("SPOTIFY_CLIENT_ID", " ")
client_secret = config.get("SPOTIFY_CLIENT_SECRET", " ")

sp = spotipy.Spotify(
    client_credentials_manager=SpotifyClientCredentials(
        client_id=client_id, client_secret=client_secret
    )
)

def get_track_info(query: str, is_uri: bool = False):
    if not is_uri:
        search_result = sp.search(query, limit=1, type="track")
        if search_result["tracks"]["items"]:
            track_item = search_result["tracks"]["items"][0]
        else:
            return False
    else:
        track_item = sp.track(query)
    return track_item

data = pd.read_csv("processed/streaming_history_tracks.csv")
data

mask_saved = data['saved'] == True
mask_uriquery = data['uri'].isnull() == False
mask_searchquery = data['uri'].isnull()

data[mask_saved]

uri_query_df = data[~mask_saved & mask_uriquery][['artist', 'track', 'album', 'uri']].drop_duplicates()
uri_query_df

search_query_df = data[~mask_saved & mask_searchquery][['artist', 'track', 'album', 'uri']].drop_duplicates()
search_query_df

search_query_df[search_query_df['artist'] == 'Arca']

search_queries = []
for row in search_query_df.itertuples():
    search_queries.append(f"{row.artist} {row.track}")

uri_queries = []
for row in uri_query_df.itertuples():
    uri_queries.append(row.uri)

print(f"""
search queries - {len(search_queries)}
uri queries - {len(uri_queries)}
""")

skipped_queries = []
tracks_info = []
tracks_audio_features = []


for query in track_progress(
    search_queries,
    total=len(search_queries),
    description="Search queries: ",
):
    track_info = get_track_info(query)
    if track_info:
        tracks_info.append(track_info)
        track_audio_features = sp.audio_features(track_info["uri"])[0]
        if track_audio_features:
            tracks_audio_features.append(track_audio_features)
    else:
        skipped_queries.append(query)

adhoc_tracks_info = []
adhoc_audio_features = []
adhoc_skipped_queries = []

for query in track_progress(
    uri_queries,
    total=len(uri_queries),
    description="URI queries: ",
):
    track_info = get_track_info(query, is_uri = True)
    if track_info:
        adhoc_tracks_info.append(track_info)
        track_audio_features = sp.audio_features(track_info["uri"])[0]
        if track_audio_features:
            adhoc_audio_features.append(track_audio_features)
    else:
        adhoc_skipped_queries.append(query)

print(f"""
skipped - {len(skipped_queries)}
track info - {len(tracks_info)}
w/ audio features - {len(tracks_audio_features)}
""")

for query in track_progress(
    ["Arca Tiro", "Arca Prada"],
    total=len(search_queries),
    description="Search queries: ",
):
    track_info = get_track_info(query)
    # if track_info:
    #     tracks_info.append(track_info)
    #     track_audio_features = sp.audio_features(track_info["uri"])[0]
    #     if track_audio_features:
    #         tracks_audio_features.append(track_audio_features)
    # else:
    #     skipped_queries.append(query)

with open("interim/tracks_info.json", "w") as file:
    json.dump(tracks_info, file, default=str)
with open("interim/tracks_audio_features.json", "w") as file:
    json.dump(tracks_audio_features, file, default=str)
