==> exploring-spotify-streaming-history/03_joiningjsons.ipynb <==
import json
from datetime import datetime as dt, timedelta

import pandas as pd
from rich import print 

data = pd.read_csv("processed/streaming_history_tracks.csv")
data['msplayed'] = pd.to_timedelta(data['msplayed'], unit='ms')
data['endtime'] = pd.to_datetime(data['endtime'])
data['played_at'] = data['endtime'] - data['msplayed']
data = data.fillna("")
data

data[data['artist'] == "Arca"]

with open("uniqueTracks.json", "r") as file:
    unique_tracks = json.load(file)

with open("interim/tracks_info.json", "r") as file:
    tracks_info = json.load(file)

# useless info for me
for track in tracks_info:
    del track['available_markets']
    del track['album']['available_markets']

tracks_info.extend([track['track'] for track in unique_tracks])

print(tracks_info[-1])

data_dict = data.to_dict(orient = "records")
print(data_dict[0])

tracks_info[1]

restructured_tracks = [{'artist': track['artists'][0]['name'], 'track': track['name'], 'album': track['album']['name'], 'uri': track['uri'], "track_artists": [artist['name'] for artist in track['artists']], 'track_info': track} for track in tracks_info]

saved_uris = [track['uri'] for track in restructured_tracks]
saved_albums = [track['album'] for track in restructured_tracks]

tracks_holder = []

for track_played in data_dict:
    current_artist = track_played['artist']
    current_track = track_played['track']
    current_album = track_played['album']
    current_uri = track_played['uri']
    current_playedat = track_played['played_at']
    if current_uri != "" and current_uri in saved_uris:
        track_data = [track for track in restructured_tracks if track['uri'] == current_uri][0]
        track_data["played_at"] = current_playedat
    elif current_album != "":
        track_data = [
            track for track in restructured_tracks
            if (
                track["artist"] == current_artist 
                and track['track'] == current_track
                and track['album'] == current_album
            )
        ][0]
    else:
        try:
            track_data = [
                track for track in restructured_tracks
                if (
                    track["artist"] == current_artist 
                    and track['track'] == current_track
                )
            ][0]
        except:
            track_data = [
                track for track in restructured_tracks
                if (
                    (current_artist in track["track_artists"])
                    and track['track'] == current_track
                )
            ]
            if len(track_data) > 0:
                track_data = track_data[0]


    track_json = {
        "artist": current_artist,
        "song": current_track,
        "album": current_album,
        "uri": current_uri,
        "played_at": current_playedat,
        "track": track_data,
    }
    tracks_holder.append(track_json)


streaming_history = [track for track in tracks_holder if type(track['track']) != list]

with open("interim/streaming_history.json", "w") as file:
    json.dump(streaming_history, file, default=str)
