==> exploring-spotify-streaming-history/04_slimming_jsons.ipynb <==
import json 
from rich import print 

import pandas as pd

with open("interim/streaming_history.json", "r") as file:
    data = json.load(file)

song = data[420]['track']['track_info']

print(song.keys())

print(song)

pd.json_normalize([track['track']['track_info'] for track in data])['album.images'].apply(len).value_counts()
